+++ go/api/response_types.go
@@ -243,3 +243,23 @@ type XPendingDetail struct {
 func CreateNilXPendingSummary() XPendingSummary {
 	return XPendingSummary{0, CreateNilStringResult(), CreateNilStringResult(), make([]ConsumerPendingMessage, 0)}
 }
+
+// XInfoGroupInfo represents a group information returned by `XInfoGroups` command.
+type XInfoGroupInfo struct {
+	// The consumer group's name.
+	Name string
+	// The number of consumers in the group.
+	Consumers int64
+	// The length of the group's Pending Entries List (PEL), which are messages that were delivered but are yet to be
+	// acknowledged.
+	Pending int64
+	// The ID of the last entry delivered to the group's consumers.
+	LastDeliveredId string
+	// The logical "read counter" of the last entry delivered to the group's consumers.
+	// Included in the response only on valkey 7.0.0 and above.
+	EntriesRead Result[int64]
+	// The number of entries in the stream that are still waiting to be delivered to the group's consumers, or a `nil` when
+	// that number can't be determined.
+	// Included in the response only on valkey 7.0.0 and above.
+	Lag Result[int64]
+}